+++ app/app.py
@@ -14,7 +14,7 @@ if "APP_DEBUG_IS_ON" in env:
     debug_on_global: Final = True
 
 from .helper.parse_configs_into_paramdict import get_param_dict
-from .helper.prepare_ml_input import Pipeline_Output, prepare_pipeline
+from .helper.prepare_ml_input import PipelineOutput, prepare_pipeline
 
 # from .model.infer_model import infer_model
 # from .model.train_model import train_model
@@ -45,7 +45,7 @@ def main(mode: APP_MODES = "train") -> None:
                 debug(item)
 
     # pipeline_objects: Pipeline_Output = prepare_pipeline(get_param_dict())
-    _: Pipeline_Output = prepare_pipeline(get_param_dict())
+    _: PipelineOutput = prepare_pipeline(get_param_dict())
 
     # _ = (
     #     train_model(**pipeline_objects)